Adding a bunch of casts.

+++ src/westmere/ip4.h
@@ -128,8 +128,8 @@ static int sse_inet_aton(const char* ipv4_string, const size_t ipv4_string_lengt
   {
     const __m128i dot = _mm_set1_epi8('.');
     const __m128i t0 = _mm_cmpeq_epi8(input, dot);
-    dotmask = _mm_movemask_epi8(t0);
-    uint16_t mask = 1 << ipv4_string_length;
+    dotmask = (uint16_t)_mm_movemask_epi8(t0);
+    uint16_t mask = (uint16_t)(1 << ipv4_string_length);
     dotmask &= mask - 1;
     dotmask |= mask;
   }
@@ -184,9 +184,9 @@ static int sse_inet_aton(const char* ipv4_string, const size_t ipv4_string_lengt
   }
   // pack and we are done!
   const __m128i t6 = _mm_packus_epi16(t5, t5);
-  uint32_t address =  _mm_cvtsi128_si32(t6);
+  uint32_t address =  (uint32_t)_mm_cvtsi128_si32(t6);
   memcpy(destination, &address, 4);
-  return ipv4_string_length - (size_t)pat[6];
+  return (int)(ipv4_string_length - (size_t)pat[6]);
 }
 
 zone_always_inline()
